03_django/CRUD_Template/requirements.txt
autopep8==1.4.4
backcall==0.1.0
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
decorator==4.4.0
Django==2.2.4
django-extensions==2.2.1
Faker==2.0.2
idna==2.8
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
parso==0.5.1
pickleshare==0.7.5
prompt-toolkit==2.0.9
pycodestyle==2.5.0
Pygments==2.4.2
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
text-unidecode==1.3
traitlets==4.3.2
urllib3==1.25.3
wcwidth==0.1.7
